Updates to two sample permuter to ensure determinism
Existing method had issues of ill-defined distributions and inconsistent thresholds run after run for smallish numbers of permutations (<1000). Using the existing randperm() approach, upwards of 40% of label permutations would be non-unique. 

Revised approach brute forces a unique list of permutations when the total number of permutations is less than 10k to ensure consistent run-to-run performance and a fully defined distribution. There is some memory overhead with this approach, but it is not inordinate for less than 10k unique label permutations.